% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{acidroxygen-package}
\alias{acidroxygen}
\alias{acidroxygen-package}
\title{acidroxygen}
\description{
Shared roxygen2 documentation for Acid Genomics packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acidroxygen.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/acidroxygen/}
  \item Report bugs at \url{https://github.com/acidgenomics/acidroxygen/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
