% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{object}{Object.}

\item{x}{Object.}

\item{y}{Object.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{assays}{\code{SimpleList}.
Count matrices, which must have matching dimensions. Counts can be passed
in as either a dense matrix (\code{matrix}) or sparse matrix (\code{sparseMatrix}).}

\item{base}{\code{integer(1)}.
Logarithm base.}

\item{cells}{\code{character}.
Cell identifiers.}

\item{censorSamples}{\code{character}.
Specify a subset of samples to censor.}

\item{class}{\code{character(1)}.
Object class.}

\item{classes}{\code{character}.
Object classes.}

\item{closed}{\code{logical(2)}.
Should the lower (1) and upper (2) bounaries be closed?}

\item{colData}{\code{DataFrame}.
Metadata describing the assay columns. For bulk RNA-seq, this data
describes the samples. For single-cell RNA-seq, this data describes the
cells.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}

\item{counts}{\code{matrix}.
Count matrix. Normalized counts are recommended.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{ensemblRelease, release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{envir}{\code{environment}.
Environment to use for assignment.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{file}{\code{character(1)}.
File path.}

\item{FUN}{\code{function}.}

\item{fun}{\code{function}.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available. Note: don't pass in UCSC build
IDs (e.g. \code{"hg38"}).}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF
(GFFv2) instead of a GFFv3 file if possible.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{lower}{\code{numeric(1)}.
Lower boundary.}

\item{metadata}{\code{list}.
Metadata.}

\item{metric}{\code{character(1)}.
Metric type, either \code{length} or \code{elements}.}

\item{msg}{\code{NULL} or \code{character(1)}.
Custom message to return.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{ntop}{\code{integer(1)}.
Number of top genes to label.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{prefilter}{\code{logical(1)}.
Apply prefiltering. Remove zero count genes.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}

\item{removeNA}{\code{logical(1)}.
Remove \code{NA} values from calculations.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{rowData}{\code{DataFrame}.
Metadata describing the assay rows, if genomic ranges are not available.
\emph{Use rowRanges (GRanges) instead, if possible}.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{samples}{\code{character}.
Sample identifiers.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{traceback}{\code{logical(1)}.
Include traceback in error message.
See \code{\link[base:traceback]{traceback()}} for details.
\code{rlang::entrace()} also works nicely and can be set in \code{.Rprofile}.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{tx2gene}{\code{Tx2Gene}.
Transcript-to-gene mappings.}

\item{unicode}{\code{logical(1)}.
Allow Unicode characters in console output.}

\item{upper}{\code{numeric(1)}.
Upper boundary.}

\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{value}{Value to assign.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}

\item{zeroPropagate}{\code{logical(1)}.
Allow propagation of zeroes.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}

\item{.yname}{Name of object defined in \code{y}.
\emph{Not intended to be used directly.}}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
