% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidRoxygen-package}
\alias{AcidRoxygen}
\alias{AcidRoxygen-package}
\title{AcidRoxygen}
\description{
Shared roxygen2 documentation for Acid Genomics packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidroxygen/}
  \item \url{https://github.com/acidgenomics/r-acidroxygen/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidroxygen/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
