% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}

\item{FUN}{\code{function}.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{asis}{\code{logical(1)}.
Set this to \code{TRUE} when using the function inside a loop or inside an R
Markdown chunk with '\code{results = "asis"}' enabled.}

\item{aspectRatio}{\code{integer(1)}.
Aspect ratio.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{assays}{\code{SimpleList}.
Count matrices, which must have matching dimensions. Counts can be passed
in as either a dense matrix (\code{matrix}) or sparse matrix (\code{sparseMatrix}).}

\item{base}{\code{integer(1)}.
Logarithm base.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{borderColor}{\code{character(1)} or \code{NULL}.
Border color.}

\item{captions}{\code{character} or \code{NULL}.
Captions.
If \code{NULL}, the names of the primary input list will be used automatically.}

\item{cells}{\code{character}.
Cell identifiers.}

\item{censorSamples}{\code{character}.
Specify a subset of samples to censor.}

\item{class}{\code{character(1)}.
Object class.}

\item{classes}{\code{character}.
Object classes.}

\item{closed}{\code{logical(2)}.
Should the lower (1) and upper (2) bounaries be closed?}

\item{colData}{\code{DataFrame}.
Metadata describing the assay columns. For bulk RNA-seq, this data
describes the samples. For single-cell RNA-seq, this data describes the
cells.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options("acid.color.discrete" = ggplot2::scale_color_viridis_d())
}\if{html}{\out{</div>}}}

\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}

\item{convertGenesToSymbols}{\code{logical(1)}.
Attempt to automatically convert gene identifiers to gene symbols.
Only applies when object contains mappings defined in \code{rowRanges}.}

\item{counts}{\code{matrix}.
Count matrix. Normalized counts are recommended.}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{acid_theme_light()} ggplot2
theme.}

\item{dims}{\code{integer}.
Vector of length 2 that denotes the columns from the reduced dimension
matrix to use for \code{centerX} and \code{centerY} column calculations. Defaults the
first and second dimensions.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{envir}{\code{environment}.
Environment.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{file}{\code{character(1)}.
File path.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options("acid.fill.discrete" = ggplot2::scale_fill_viridis_d())
}\if{html}{\out{</div>}}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{fun}{\code{function}.}

\item{geneToSymbol}{\code{GeneToSymbol}.
Gene-to-symbol mappings. Must contain \code{geneId} and \code{geneName} columns. See
\code{GeneToSymbol} for more information.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file.
Generally, we recommend using GTF (GFFv2) instead of GFFv3.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{ignoreTxVersion}{\code{logical(1)}. Don't the include the transcript version
in the identifier.}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{lower}{\code{numeric(1)}.
Lower boundary.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{metadata}{\code{list}.
Metadata.}

\item{metric}{\code{character(1)}.
Metric type, either \code{length} or \code{elements}.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per feature (i.e. gene).}

\item{msg}{\code{NULL} or \code{character(1)}.
Custom message to return.}

\item{nMax}{\code{integer(1)}. or \code{Inf}.
Maximum number (i.e. of lines) to include.
Use \code{Inf} to not apply a limit.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{ntop}{\code{integer(1)}.
Number of top features to label.}

\item{nullOk}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{object}{Object.}

\item{ordered}{\code{logical(1)}.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointScalar}{\code{integer(1)}.
Default point size for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{prefilter}{\code{logical(1)}.
Apply prefiltering. Remove zero count genes.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{projectDir}{\code{character(1)}.
Project directory path.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{removeNa}{\code{logical(1)}.
Remove \code{NA} values from calculations.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{rowData}{\code{DataFrame}.
Metadata describing the assay rows, if genomic ranges are not available.
\emph{Use \code{rowRanges} (\code{GenomicRanges}) instead, if possible}.}

\item{rowRanges}{\code{GenomicRanges} or \code{GenomicRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{samples}{\code{character}.
Sample identifiers.}

\item{sigPointColor}{\code{character}.
Color names for labeling upregulated and downregulated genes. Also supports
a character string for labeling DEGs with the same color, regardless of
direction.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{subtitle}{\code{character(1)}.
Subtitle.}

\item{tabset}{\code{logical(1)}.
Include tabset marker.}

\item{text}{\code{character(1)}.
Text.}

\item{title}{\code{character(1)}.
Title.}

\item{traceback}{\code{logical(1)}.
Include traceback in error message.
See \code{traceback()} for details.
\code{rlang::entrace()} also works nicely and can be set in \code{.Rprofile}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}\if{html}{\out{</div>}}}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{txToGene}{\code{TxToGene}.
Transcript-to-gene mappings.}

\item{unicode}{\code{logical(1)}.
Allow Unicode characters in console output.}

\item{uploadDir}{\code{character(1)}.
Final upload directory path.}

\item{upper}{\code{numeric(1)}.
Upper boundary.}

\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{value}{Value to assign.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}

\item{x}{Object.}

\item{y}{Object.}

\item{yaml}{\code{list}.
YAML metadata.}

\item{zeroPropagate}{\code{logical(1)}.
Allow propagation of zeroes.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}

\item{.yname}{Name of object defined in \code{y}.
\emph{Not intended to be used directly.}}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
